Split up DefaultApplication
