+++ Stormpath.SDK/Stormpath.SDK/Impl/Application/DefaultApplication.Collections.cs
@@ -1,0 +1,33 @@
+﻿// <copyright file="DefaultApplication.Collections.cs" company="Stormpath, Inc.">
+// Copyright (c) 2015 Stormpath, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using Stormpath.SDK.Account;
+using Stormpath.SDK.Application;
+using Stormpath.SDK.Group;
+using Stormpath.SDK.Impl.Linq;
+using Stormpath.SDK.Linq;
+
+namespace Stormpath.SDK.Impl.Application
+{
+    internal sealed partial class DefaultApplication
+    {
+        IAsyncQueryable<IAccount> IApplication.GetAccounts()
+            => new CollectionResourceQueryable<IAccount>(this.Accounts.Href, this.GetInternalAsyncDataStore());
+
+        IAsyncQueryable<IGroup> IApplication.GetGroups()
+            => new CollectionResourceQueryable<IGroup>(this.Groups.Href, this.GetInternalAsyncDataStore());
+    }
+}
