+++ Stormpath.SDK/Stormpath.SDK/Impl/Application/DefaultApplication.Resources.cs
@@ -1,0 +1,34 @@
+﻿// <copyright file="DefaultApplication.Resources.cs" company="Stormpath, Inc.">
+// Copyright (c) 2015 Stormpath, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using System.Threading;
+using System.Threading.Tasks;
+using Stormpath.SDK.Application;
+using Stormpath.SDK.Impl.Provider;
+using Stormpath.SDK.Provider;
+using Stormpath.SDK.Tenant;
+
+namespace Stormpath.SDK.Impl.Application
+{
+    internal sealed partial class DefaultApplication
+    {
+        Task<IProviderAccountResult> IApplication.GetAccountAsync(IProviderAccountRequest request, CancellationToken cancellationToken)
+            => new ProviderAccountResolver(this.GetInternalAsyncDataStore()).ResolveProviderAccountAsync(this.AsInterface.Href, request, cancellationToken);
+
+        Task<ITenant> IApplication.GetTenantAsync(CancellationToken cancellationToken)
+            => this.GetTenantAsync(this.Tenant.Href, cancellationToken);
+    }
+}
