+++ Stormpath.SDK/Stormpath.SDK/Impl/Application/DefaultApplication.cs
@@ -118,107 +118,6 @@ namespace Stormpath.SDK.Impl.Application
             return this;
         }
 
-        Task<IAuthenticationResult> IApplication.AuthenticateAccountAsync(IAuthenticationRequest request, CancellationToken cancellationToken)
-        {
-            var dispatcher = new AuthenticationRequestDispatcher();
-
-            return dispatcher.AuthenticateAsync(this.GetInternalAsyncDataStore(), this, request, null, cancellationToken);
-        }
-
-        IAuthenticationResult IApplicationSync.AuthenticateAccount(IAuthenticationRequest request)
-        {
-            var dispatcher = new AuthenticationRequestDispatcher();
-
-            return dispatcher.Authenticate(this.GetInternalSyncDataStore(), this, request, null);
-        }
-
-        Task<IAuthenticationResult> IApplication.AuthenticateAccountAsync(IAuthenticationRequest request, Action<IRetrievalOptions<IAuthenticationResult>> responseOptions, CancellationToken cancellationToken)
-        {
-            var options = new DefaultRetrievalOptions<IAuthenticationResult>();
-            responseOptions(options);
-
-            var dispatcher = new AuthenticationRequestDispatcher();
-
-            return dispatcher.AuthenticateAsync(this.GetInternalAsyncDataStore(), this, request, options, cancellationToken);
-        }
-
-        IAuthenticationResult IApplicationSync.AuthenticateAccount(IAuthenticationRequest request, Action<IRetrievalOptions<IAuthenticationResult>> responseOptions)
-        {
-            var options = new DefaultRetrievalOptions<IAuthenticationResult>();
-            responseOptions(options);
-
-            var dispatcher = new AuthenticationRequestDispatcher();
-
-            return dispatcher.Authenticate(this.GetInternalSyncDataStore(), this, request, options);
-        }
-
-        Task<IAuthenticationResult> IApplication.AuthenticateAccountAsync(string username, string password, CancellationToken cancellationToken)
-        {
-            var request = new UsernamePasswordRequest(username, password) as IAuthenticationRequest;
-
-            return this.AsInterface.AuthenticateAccountAsync(request, cancellationToken);
-        }
-
-        IAuthenticationResult IApplicationSync.AuthenticateAccount(string username, string password)
-        {
-            var request = new UsernamePasswordRequest(username, password) as IAuthenticationRequest;
-
-            return this.AuthenticateAccount(request);
-        }
-
-        async Task<bool> IApplication.TryAuthenticateAccountAsync(string username, string password, CancellationToken cancellationToken)
-        {
-            try
-            {
-                var loginResult = await this.AsInterface.AuthenticateAccountAsync(username, password, cancellationToken).ConfigureAwait(false);
-                return true;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        bool IApplicationSync.TryAuthenticateAccount(string username, string password)
-        {
-            try
-            {
-                var loginResult = this.AuthenticateAccount(username, password);
-                return true;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        Task<IAccount> IAccountCreationActions.CreateAccountAsync(IAccount account, Action<AccountCreationOptionsBuilder> creationOptionsAction, CancellationToken cancellationToken)
-            => AccountCreationActionsShared.CreateAccountAsync(this.GetInternalAsyncDataStore(), this.Accounts.Href, account, creationOptionsAction, cancellationToken);
-
-        IAccount IAccountCreationActionsSync.CreateAccount(IAccount account, Action<AccountCreationOptionsBuilder> creationOptionsAction)
-            => AccountCreationActionsShared.CreateAccount(this.GetInternalSyncDataStore(), this.Accounts.Href, account, creationOptionsAction);
-
-        Task<IAccount> IAccountCreationActions.CreateAccountAsync(IAccount account, IAccountCreationOptions creationOptions, CancellationToken cancellationToken)
-             => AccountCreationActionsShared.CreateAccountAsync(this.GetInternalAsyncDataStore(), this.Accounts.Href, account, creationOptions, cancellationToken);
-
-        IAccount IAccountCreationActionsSync.CreateAccount(IAccount account, IAccountCreationOptions creationOptions)
-             => AccountCreationActionsShared.CreateAccount(this.GetInternalSyncDataStore(), this.Accounts.Href, account, creationOptions);
-
-        Task<IAccount> IAccountCreationActions.CreateAccountAsync(IAccount account, CancellationToken cancellationToken)
-             => AccountCreationActionsShared.CreateAccountAsync(this.GetInternalAsyncDataStore(), this.Accounts.Href, account, cancellationToken);
-
-        IAccount IAccountCreationActionsSync.CreateAccount(IAccount account)
-             => AccountCreationActionsShared.CreateAccount(this.GetInternalSyncDataStore(), this.Accounts.Href, account);
-
-        Task<IAccount> IAccountCreationActions.CreateAccountAsync(string givenName, string surname, string email, string password, object customData, CancellationToken cancellationToken)
-            => AccountCreationActionsShared.CreateAccountAsync(this.GetInternalAsyncDataStore(), this.Accounts.Href, givenName, surname, email, password, customData, cancellationToken);
-
-        Task<IAccount> IAccountCreationActions.CreateAccountAsync(string givenName, string surname, string email, string password, CancellationToken cancellationToken)
-            => AccountCreationActionsShared.CreateAccountAsync(this.GetInternalAsyncDataStore(), this.Accounts.Href, givenName, surname, email, password, null, cancellationToken);
-
-        IAccount IAccountCreationActionsSync.CreateAccount(string givenName, string surname, string email, string password, object customData)
-            => AccountCreationActionsShared.CreateAccount(this.GetInternalSyncDataStore(), this.Accounts.Href, givenName, surname, email, password, customData);
-
         Task<bool> IDeletable.DeleteAsync(CancellationToken cancellationToken)
             => this.GetInternalAsyncDataStore().DeleteAsync(this, cancellationToken);
 
@@ -236,73 +135,5 @@ namespace Stormpath.SDK.Impl.Application
 
         IApplication ISaveableWithOptionsSync<IApplication>.Save(Action<IRetrievalOptions<IApplication>> options)
              => this.Save(options);
-
-        Task IApplication.SendVerificationEmailAsync(string usernameOrEmail, CancellationToken cancellationToken)
-            => this.AsInterface.SendVerificationEmailAsync(request => request.Login = usernameOrEmail, cancellationToken);
-
-        Task IApplication.SendVerificationEmailAsync(Action<EmailVerificationRequestBuilder> requestBuilderAction, CancellationToken cancellationToken)
-        {
-            var builder = new EmailVerificationRequestBuilder(this.GetInternalAsyncDataStore());
-            requestBuilderAction(builder);
-
-            if (string.IsNullOrEmpty(builder.Login))
-                throw new ArgumentNullException(nameof(builder.Login));
-
-            var href = $"{(this as IResource).Href}/verificationEmails";
-
-            return this.GetInternalAsyncDataStore().CreateAsync(href, builder.Build(), cancellationToken);
-        }
-
-        void IApplicationSync.SendVerificationEmail(string usernameOrEmail)
-            => this.SendVerificationEmail(request => request.Login = usernameOrEmail);
-
-        void IApplicationSync.SendVerificationEmail(Action<EmailVerificationRequestBuilder> requestBuilderAction)
-        {
-            var builder = new EmailVerificationRequestBuilder(this.GetInternalAsyncDataStore());
-            requestBuilderAction(builder);
-
-            if (string.IsNullOrEmpty(builder.Login))
-                throw new ArgumentNullException(nameof(builder.Login));
-
-            var href = $"{(this as IResource).Href}/verificationEmails";
-
-            this.GetInternalSyncDataStore().Create(href, builder.Build());
-        }
-
-        Task<IGroup> IGroupCreationActions.CreateGroupAsync(IGroup group, CancellationToken cancellationToken)
-            => GroupCreationActionsShared.CreateGroupAsync(this.GetInternalAsyncDataStore(), this.Groups.Href, group, cancellationToken);
-
-        IGroup IGroupCreationActionsSync.CreateGroup(IGroup group)
-            => GroupCreationActionsShared.CreateGroup(this.GetInternalSyncDataStore(), this.Groups.Href, group);
-
-        Task<IGroup> IGroupCreationActions.CreateGroupAsync(IGroup group, Action<GroupCreationOptionsBuilder> creationOptionsAction, CancellationToken cancellationToken)
-            => GroupCreationActionsShared.CreateGroupAsync(this.GetInternalAsyncDataStore(), this.Groups.Href, group, creationOptionsAction, cancellationToken);
-
-        IGroup IGroupCreationActionsSync.CreateGroup(IGroup group, Action<GroupCreationOptionsBuilder> creationOptionsAction)
-            => GroupCreationActionsShared.CreateGroup(this.GetInternalSyncDataStore(), this.Groups.Href, group, creationOptionsAction);
-
-        Task<IGroup> IGroupCreationActions.CreateGroupAsync(IGroup group, IGroupCreationOptions creationOptions, CancellationToken cancellationToken)
-            => GroupCreationActionsShared.CreateGroupAsync(this.GetInternalAsyncDataStore(), this.Groups.Href, group, creationOptions, cancellationToken);
-
-        IGroup IGroupCreationActionsSync.CreateGroup(IGroup group, IGroupCreationOptions creationOptions)
-            => GroupCreationActionsShared.CreateGroup(this.GetInternalSyncDataStore(), this.Groups.Href, group, creationOptions);
-
-        Task<IProviderAccountResult> IApplication.GetAccountAsync(IProviderAccountRequest request, CancellationToken cancellationToken)
-            => new ProviderAccountResolver(this.GetInternalAsyncDataStore()).ResolveProviderAccountAsync(this.AsInterface.Href, request, cancellationToken);
-
-        IProviderAccountResult IApplicationSync.GetAccount(IProviderAccountRequest request)
-            => new ProviderAccountResolver(this.GetInternalAsyncDataStore()).ResolveProviderAccount(this.AsInterface.Href, request);
-
-        Task<ITenant> IApplication.GetTenantAsync(CancellationToken cancellationToken)
-            => this.GetTenantAsync(this.Tenant.Href, cancellationToken);
-
-        ITenant IApplicationSync.GetTenant()
-            => this.GetTenant(this.Tenant.Href);
-
-        IAsyncQueryable<IAccount> IApplication.GetAccounts()
-             => new CollectionResourceQueryable<IAccount>(this.Accounts.Href, this.GetInternalAsyncDataStore());
-
-        IAsyncQueryable<IGroup> IApplication.GetGroups()
-            => new CollectionResourceQueryable<IGroup>(this.Groups.Href, this.GetInternalAsyncDataStore());
     }
 }
